Move entity manager to Map

+++ Constants.cpp
@@ -3,8 +3,8 @@
 
 #include "Constants.h"
 
-sf::Vector2f Constants::mapCenterAsPixels() {
-    return sf::Vector2f(Constants::MAP_WIDTH * Constants::TILE_WIDTH / 2,
+sf::Vector2i Constants::mapCenterAsPixels() {
+    return sf::Vector2i(Constants::MAP_WIDTH * Constants::TILE_WIDTH / 2,
                         Constants::MAP_HEIGHT * Constants::TILE_HEIGHT / 2);
 }
